Re-write merge-compatible-upgrades for github-script@v5
Dependabot automatically upgraded to actions/github-script@v5, which
contained breaking changes. The changes were not automatically detected
since the breaking changes were inside a mock. Preventing this may be
possible, but it is not obvious how since they are not npm packages.
Maybe by cloning the repo? A problem for another day.